refactor: move header functions to headers.ts, add retry-after seconds, remove object spreads
- Move filterResponseHeaders, buildRetryHeaders, and header constants to headers.ts
- Add retry-after (seconds) alongside retry-after-ms for SDK compatibility
- Derive missing counterpart: retry-after-ms → retry-after (ceil) and vice versa
- Replace object spread in prepareResponseInit with mutation
