+++ test/e2e/chat-completions/chat-completions.test.ts
@@ -6,7 +6,10 @@ import type { ChatCompletionMessageFunctionToolCall } from "openai/resources/cha
 import { gptOss120b } from "../../../src/models/openai";
 import { BEDROCK_ACCESS_KEY_ID, BEDROCK_SECRET_ACCESS_KEY } from "../shared/server";
 import { createBedrockTestServer, type TestServer } from "../shared/server";
-import { CHAT_CALCULATOR_TOOL as CALCULATOR_TOOL, CHAT_WEATHER_TOOL as WEATHER_TOOL } from "../shared/tools";
+import {
+  CHAT_CALCULATOR_TOOL as CALCULATOR_TOOL,
+  CHAT_WEATHER_TOOL as WEATHER_TOOL,
+} from "../shared/tools";
 
 // ---------------------------------------------------------------------------
 // Environment
